Starting Move Rework
Old method was using a 32 bit value to hold the 8 directions and their
longest possible distance traveled, plus whether the final value was a
capture or not. This was good for saving space, but would have just
doubled the amount of work being done every time moves are calculated:
loop through the board in every direction from the piece to find the
closest piece to set the move distance, and then when creating the
possible moves you would have to loop through that value again to create
all of the valid moves.

Instead, (2) 64 bit values will be used, one with bits set for valid
moves, the other with bits set for moves that would be captures. This
will give all of the possible moves without having to reloop through
0-maxDistance