BIG Commit
Old laptop broke, and had to transfer this over to a new laptop, which did not have git setup. Should be committing like normal from now on. The program appears to be in a generally working state right now: bitFields seem to line up, moves seem to be correctly calculated and carried out.

Things that need to be done:
-Pawn captures, including en passant
-Check for check(mate)
-Check for castling and implementing the move