Added Board Function for Active Play
Instead of hoping situations just arise, or setting them up in code,
added a function that will print all of the available moves, and let you
choose the move to make. En Passant appears to be correctly identified,
I still need to check to make sure that there aren't false positives.

I' already seeing there is an issue with the En Passant check, where
there's no check to make sure the last move is in the correct column.
That should be added in the next update.