Moved undoMove to Just memcpy
For now while I figure out that undoMove (still needed to be able to
undo a move when playing, by the player), I changed to just copy over
the current pieceMap, make a move, check for check, and then just copy
it back. Still facing an issue with checks, it appears the king's spot
is not staying where it should.