Pulled Code Out of moveLeadsToCheck
Made into its own function, board_oppositeKingPressured. Writing out
what is need for looking for moves that lead to check (for both
attacking and preventing invalid moves) and checkmate, it seems like it
will be better to have that function rather than relying on making a
move to check for check. There is also a segmentation fault that is
happening with certain move combinations, appears to be in
getMovesForCurrentSide, need to look into it more.