Added Functionality to Check Function For/Against
Until now, I was only using that function to see if a move would put the
opposite player into check. I also needed a solution for seeing if a
move would lead to the same player being in check, AKA being an invalid
move. Turns out, the only difference between the two is whether or not
to flip the current player after making the move.

If you are looking to see if a move would put your opponent check, make
the move (which flips the current player), and flip it back before
seeing if the opposite king is in check. If you are looking to see if a
move would put yourself in check, and be invalid, make the move, and
don't flip the current player again so that the function will check if
the opponent can attack your king.