Adjusted makeMove, Added Check for AgainstMe Check
makeMove needed a bit of an overhaul because of how confusing it was,
and because now there is an opportunity for the move to not be valid.
Moved a lot of the generic things (src/dstRow/Col, whiteMove, and
pieceType) to the top since they would be the same for every move, and
only focused on what was different about moves when they happened.

Added the check to see if a move would let the opponent attack your
king, not adding the move if so. From a short little test, it worked,
will need to do further testing though.

I believe, and this is just speculation, that this now opens up an easy
way to check for checkmate (for the current player, not predicting
early, but it may be able to help). If the current player can't come up
with any moves, that must mean they can't move to a position without
allowing an attack on their king, so if they are currently in check they
lose.