Setting inCheck Boolean to false in makeMove
Function before only held the code that was setting it to true, now it
will set it back to false if the move is not leading to check. I believe
the only time I'll actually need to check this variable is when checking
if a move would be stalemate or checkmate based on whether the opponent
is in check. May also be useful for a visual indicator that the player
is in check.