Starting Check/Stale Mate Checks
I believe the thought process and code are fairly sound, but seg fault
does happen when running. Almost certainly the fact that leadsToCheck,
called with getMoveForCurrentSide, is then calling
getMoveForCurrentSize, which will then call leadsToCheck, forever. Need
to figure out way to prevent this, dumbest is probably just adding a
boolean parameter that changes getMovesForCurrentSide's behaviour.

Smarter is probably to have seperate functions for performing different
actions: one function that only gets the possible moves, not taking into
account their ability for checks, and then a function that takes in a
*Move and removes/modifies moves as needed.