Fixed leadsTocheck Error with Early Exit
If the move would lead to the current player being checked, the function
exits early, which is great. What was not great was not reverting the
board back to what it was, so having it be totally messed up for the
next move check. Updated, appears to be working pretty great now.