Starting Game Runner
Getting the skeleton for running games. Created a function that will try
to make a decision on what move to make to allow for easy
changing/testing, will probably need to change the parameters as the
options get more and more complex. For now, working as just the same
user-input lead game that I had before.