Added wht/black BoardPressure BitFields
Updated every turn, and then show where each piece can attack the next
turn. Can probably use this to streamline the check function, for now I
am going to be updating castling so it actually works as intended,
making it so the king can't move through a square being attacked by the
opponent.